Update and partial completion with username/api/leaderboard handler
general update and orginization with classes, to finish my API updating the leaderboard I need the score system build and a datastructure to hold the information that will be sent out to the server. I also included my backend files I created using cloud9. I have my webserver running apache2 and it just holds the small amount of php code that will parse the data and recieve it from the user. I also added lots of comments to explain changes as I went, throughout the code I have a few placeholders for testing purposes which will be removed when we have legit data to send to the website.